+++ tools/dvd-spi-stubs/sage/DVDStreamTransformProvider.java
@@ -1,0 +1,8 @@
+package sage;
+import java.io.IOException;
+public interface DVDStreamTransformProvider {
+ String getTransportId();
+ String getOutputFormat();
+ boolean isAvailable();
+ DVDStreamTransform open(DVDStreamTransformRequest request) throws IOException;
+}